moved 404 route to the bottom; moved GuestHomePage to pages folder;  removed About; added common.json to en translations
