+++ src/constants/translations/en/common.json
@@ -1,0 +1,4 @@
+{
+  "title": "space2",
+  "about": "About us"
+}
