+++ src/constants/translations/en/index.js
@@ -1,5 +1,6 @@
+import common from './common.json'
 import errorPage from './errorPage.json'
 import button from './button.json'
 
 // eslint-disable-next-line import/no-anonymous-default-export
-export default { translations: { errorPage, button } }
+export default { translations: { common, errorPage, button } }
