+++ src/containers/layout/AppMain.js
@@ -1,17 +1,17 @@
 import { Route, Routes } from 'react-router-dom'
+
 import { routes } from '~/constants/routes'
-import Home from '~/pages/home/Home'
+import Example from '~/pages/home/Home'
+import GuestHomePage from '~/pages/guest-home-page/GuestHomePage'
 import NotFound from '~/pages/error/NotFound'
-import About from '~/pages/about/About'
-import GuestHomePage from '~/components/GuestHomePage/GuestHomePage'
+
 
 const AppMain = () => {
   return (
     <Routes>
-      <Route element={ <NotFound /> } path="*" />
       <Route element={ <GuestHomePage /> } path={ routes.home } />
-      { /* <Route element={ <Home /> } path={ routes.home } /> */ }
-      <Route element={ <About /> } name="home" path={ routes.about } />
+      <Route element={ <Example /> } name="home" path={ routes.about } />
+      <Route element={ <NotFound /> } path="*" />
     </Routes>
   )
 }
