+++ src/pages/error/NotFound.js
@@ -1,8 +1,11 @@
 import React from 'react'
 
+
 const NotFound = () => {
   return (
-    <div>404 NotFound</div>
+    <div>
+      404 NotFound
+    </div>
   )
 }
