+++ src/pages/guest-home-page/GuestHomePage.js
@@ -1,6 +1,6 @@
 import React from 'react'
 
-/* path to this file?? */
+
 const GuestHomePage = () => {
   return (
     <div>
